añadido comentarios y reestructuración
