+++ src/pruebas/mostrarprueba.ui
