+++ src/pruebas/pruebadiagnostica.cpp
@@ -1,21 +1,6 @@
-#include "pruebadiagnostica.h"
+#include <pruebas/pruebadiagnostica.h>
 
-PruebaDiagnostica::PruebaDiagnostica()
-{
-
-}
-
-PCR::PCR()
-{
-
-}
-
-Antigenos::Antigenos()
-{
-
-}
-
-string PruebaDiagnostica::getdni()
+std::string PruebaDiagnostica::getdni()
 {
     return dni;
 }
@@ -25,7 +10,7 @@ QDate PruebaDiagnostica::getFecha()
     return fechaMuestra;
 }
 
-void PruebaDiagnostica::setdni(string dni_in)
+void PruebaDiagnostica::setdni(std::string dni_in)
 {
     dni=dni_in;
 }
@@ -35,12 +20,12 @@ void PruebaDiagnostica::setFecha(int dia, int mes, int anyo)
     fechaMuestra=QDate(anyo,mes,dia);
 }
 
-string Antigenos::getnombre()
+std::string Antigenos::getnombre()
 {
     return "Antigenos";
 }
 
-string PCR::getnombre()
+std::string PCR::getnombre()
 {
     return "PCR";
 }
